Update name mangling description to explicitly condone the practice of using a period-separated suffix for symbol variants.
The intent is to provide a mechanism whereby implementations can provide
multiple definitions of a symbol (eg, a constant-propagated version of a
function, the residual after partial inlining or dead argument
elimination, a resumption function for a coroutine, ...) that are named
after the original in backtraces, debuggers, and so on.

Fixes #35.