Fix description of empty classes to handle unnamed bit-fields of nonzero length.
Unnamed bit-fields are not non-static data members, so the definition of "empty class" disregarded them. Clang follows the letter of this rule; GCC and ICC follow the intent, which is that unnamed bit-fields of non-zero length render a class object non-empty.

The C++ standard has the same bug, which is currently in the process of being corrected by wg21.link/lwg2358.