Don't incorporate cookie alignment into new[] allocations
Per discussion with @zygoloid, the standard does not permit cookie alignment to be factored into allocations, either by changing the resolution of operator new[] or changing the `std::align_val_t` argument.  We could ask for this to be changed, but it's not clear that there's any reasonable solution for class-specific allocators that don't take an explicit alignment, which it seems reasonable to say should be allowed to return something based on the alignment of the class.  Given that, the only reasonable decision is to not always align the cookie.  We should still be able to assume it's aligned in common cases.